Change to be exposed plugin method. Closes #3. - Removed all routes the plugin adds - Clone process.env - Updated README with changes
