+++ lib/index.js
@@ -1,9 +1,5 @@
 'use strict';
 var Hoek = require('hoek');
-var defaults = {
-  path: '/hapi-setup',
-  dataPath: '/hapi-setup/data'
-};
 
 function constructResponse (server) {
   var runtime = {
@@ -12,13 +8,13 @@ function constructResponse (server) {
     argv: Hoek.clone(process.argv),
     execArgv: Hoek.clone(process.execArgv),
     cwd: process.cwd(),
-    // This should probably be a clone(), but Hoek is currently breaking
-    // when working with process.env on 0.10
-    env: process.env
+    env: Hoek.clone(process.env)
   };
   var connections = server.connections.map(function (connection) {
     var plugins = {};
 
+    // Don't use _registrations.
+    // related issue: https://github.com/hapijs/hapi/issues/2777
     Object.keys(connection._registrations).forEach(function (name) {
       var registration = connection._registrations[name];
 
@@ -56,34 +52,9 @@ function constructResponse (server) {
 }
 
 module.exports.register = function (server, options, next) {
-  options = Hoek.applyToDefaults(defaults, options, true, false);
-  server.route([
-    {
-      method: 'GET',
-      path: options.path,
-      config: {
-        description: 'Display server configuration information as HTML',
-        handler: function (request, reply) {
-          // var data = constructResponse(request.server);
-          var html = '<!DOCTYPE html><html><body>Coming Soon</body></html>';
-
-          reply(html);
-        }
-      }
-    },
-    {
-      method: 'GET',
-      path: options.dataPath,
-      config: {
-        description: 'Display server configuration information as JSON',
-        handler: function (request, reply) {
-          var data = constructResponse(request.server);
-
-          reply(data);
-        }
-      }
-    }
-  ]);
+  server.expose('setup', function () {
+    return constructResponse(server);
+  });
 
   next();
 };
